requirements.txt

Flask>=3.0,<3.1
google-generativeai>=0.6,<0.9
gunicorn>=21,<23
Pillow>=10,<11
google-cloud-texttospeech==2.16.3
edge-tts
python-dotenv==1.0.0

google-generativeai
edge-tts
Pillow
python-dotenv
